feat(news): implement article submission with security enhancements
- Add slugify and xss dependencies for slug generation and input sanitization
- Move form submission logic to serverAction service
- Remove manual slug input field from form
- Implement addArticle function in news service
